requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.38.0
anyio==4.6.2.post1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
copilotkit==0.1.25
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.4
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-anthropic==0.2.4
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langgraph==0.2.44
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.139
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
openai==1.53.0
orjson==3.10.11
packaging==24.1
partialjson==0.0.8
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tqdm==4.66.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
